Adds UI support for custom dataset properties (#1905)
